// InventoryManagementUI/src/openapi/services/model/items.ts
/**
 * Seyir Füzeleri Sistem Mühendisliği Envanter Uygulaması API
 * NodeJS - ExpressJS API Docs / Fevzi TEMİZ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TypeOfItems } from './typeOfItems';
import { Brands } from './brands';


export interface Items { 
    id: number;
    model: string;
    serialNumber: string;
    eRegistryNumber: string;
    description: string;
    brand: Brands;
    typeOfItem: TypeOfItems;
}